refactor: Merge Flash optimizations into RingDilatedAttentionV2Collective
- Integrated all Flash Attention functionality from V2Flash into V2Collective
- Added use_flash_attention and flash_chunk_size parameters
- Automatic backend selection (xformers, SDPA, Flash Attention)
- Memory-efficient chunked processing for very long sequences
- Removed RingDilatedAttentionV2Flash class (no longer needed)
- Updated RingMultiheadDilatedAttention to use enhanced V2Collective

Benefits:
- Single implementation reduces confusion and maintenance
- All users get Flash optimizations by default
- Backward compatible with use_flash_attention parameter
- Cleaner codebase with less duplication
